Remove drmCheckModesettingSupported and kernel module loading, on Linux
The former of these is a UMS artefact which gives incorrect and
misleading promise whether KMS is supported. Not to mention that
AMDGPU is a only KMS driver.

In a similar fashion xf86LoadKernelModule() is a relic of the times,
where platforms had no scheme of detecting and loading the appropriate
kernel module.

Notes:
 - Since there is no reply from Robert the code is still around, behind
a FreeBSD guard.
 - If FreeBSD still needs this they should look and fix it ASAP, as:
   - wayland itself or compositors do _not_ load kernel modules
   - the kernel module should be loaded early to control the clocks/fan,
hence temperature of the card

v2: Keep the code as FreeBSD only, add 'Notes' in the commit message.

Cc: Robert Millan <rmh@freebsd.org>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
Reviewed-by: Michel Dänzer <michel.daenzer@amd.com>
Acked-by: Alex Deucher <alexander.deucher@amd.com>