Check last flip window instead of screen root before flipping
A significant amount of time can pass between the X call into
check_flip for a window and when amdgpu_present_flip actually occurs.
To ensure that flipping is still possible there was an additional check
performed on screen->root in amdgpu_present_flip - but what should
be checked instead is the window itself. This only really worked before
because X ensures that the window has the same dimensions as the screen
to allow for present extension flipping.

This patch tracks the flip window between calls to check_flip and flip
and uses that window instead of screen->root.

Signed-off-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>